pull out helper methods from rest api integration tests into a util file.
