+++ ksql-rest-app/src/test/java/io/confluent/ksql/rest/integration/ClusterTerminationTest.java
@@ -22,20 +22,11 @@ import com.google.common.collect.ImmutableList;
 import io.confluent.common.utils.IntegrationTest;
 import io.confluent.ksql.integration.IntegrationTestHarness;
 import io.confluent.ksql.rest.client.KsqlRestClient;
-import io.confluent.ksql.rest.client.RestResponse;
 import io.confluent.ksql.rest.entity.ClusterTerminateRequest;
-import io.confluent.ksql.rest.entity.CommandStatus;
-import io.confluent.ksql.rest.entity.CommandStatus.Status;
-import io.confluent.ksql.rest.entity.CommandStatusEntity;
-import io.confluent.ksql.rest.entity.KsqlEntity;
-import io.confluent.ksql.rest.entity.KsqlEntityList;
-import io.confluent.ksql.rest.entity.KsqlRequest;
 import io.confluent.ksql.rest.server.TestKsqlRestApp;
 import io.confluent.ksql.util.KsqlConstants;
 import io.confluent.ksql.util.PageViewDataProvider;
-import java.util.Collections;
 import java.util.List;
-import java.util.stream.Collectors;
 import javax.ws.rs.client.Client;
 import javax.ws.rs.client.Entity;
 import javax.ws.rs.core.MediaType;
@@ -75,7 +66,7 @@ public class ClusterTerminationTest {
   public static void setUpClass() {
     TEST_HARNESS.ensureTopics(PAGE_VIEW_TOPIC);
 
-    createStreams(PAGE_VIEW_STREAM);
+    RestIntegrationTestUtil.createStreams(REST_APP, PAGE_VIEW_STREAM, PAGE_VIEW_TOPIC);
   }
 
   @Before
@@ -91,9 +82,12 @@ public class ClusterTerminationTest {
   @Test
   public void shouldCleanUpSinkTopicsAndSchemasDuringClusterTermination() throws Exception {
     // Given:
-    makeKsqlRequest("CREATE STREAM " + SINK_STREAM
-        + " WITH (kafka_topic='" + SINK_TOPIC + "',value_format='avro')"
-        + " AS SELECT * FROM " + PAGE_VIEW_STREAM + ";"
+    RestIntegrationTestUtil.makeKsqlRequest(
+        REST_APP,
+        restClient,
+        "CREATE STREAM " + SINK_STREAM
+            + " WITH (kafka_topic='" + SINK_TOPIC + "',value_format='avro')"
+            + " AS SELECT * FROM " + PAGE_VIEW_STREAM + ";"
     );
 
     TEST_HARNESS.waitForTopicsToBePresent(SINK_TOPIC);
@@ -112,77 +106,6 @@ public class ClusterTerminationTest {
     TEST_HARNESS.waitForSubjectToBeAbsent(SINK_TOPIC + KsqlConstants.SCHEMA_REGISTRY_VALUE_SUFFIX);
   }
 
-  private static List<KsqlEntity> awaitResults(final List<KsqlEntity> pending) {
-    try (final KsqlRestClient ksqlRestClient = REST_APP.buildKsqlClient()) {
-      return pending.stream()
-          .map(e -> awaitResult(e, ksqlRestClient))
-          .collect(Collectors.toList());
-    }
-  }
-
-  private static KsqlEntity awaitResult(
-      final KsqlEntity e,
-      final KsqlRestClient ksqlRestClient
-  ) {
-    if (!(e instanceof CommandStatusEntity)) {
-      return e;
-    }
-
-    CommandStatusEntity cse = (CommandStatusEntity) e;
-    final String commandId = cse.getCommandId().toString();
-
-    while (cse.getCommandStatus().getStatus() != Status.ERROR
-        && cse.getCommandStatus().getStatus() != Status.SUCCESS) {
-
-      final RestResponse<CommandStatus> res = ksqlRestClient.makeStatusRequest(commandId);
-
-      throwOnError(res);
-
-      cse = new CommandStatusEntity(
-          cse.getStatementText(),
-          cse.getCommandId(),
-          res.getResponse(),
-          cse.getCommandSequenceNumber()
-      );
-    }
-
-    return cse;
-  }
-
-  @SuppressWarnings("unchecked")
-  private List<KsqlEntity> makeKsqlRequest(final String sql) {
-    final RestResponse<KsqlEntityList> res = restClient.makeKsqlRequest(sql);
-
-    throwOnError(res);
-
-    return awaitResults(res.getResponse());
-  }
-
-  private static void throwOnError(final RestResponse<?> res) {
-    if (res.isErroneous()) {
-      throw new AssertionError("Failed to await result."
-          + "msg: " + res.getErrorMessage());
-    }
-  }
-
-  private static void createStreams(final String streamName) {
-    final Client client = TestKsqlRestApp.buildClient();
-
-    try (final Response response = client
-        .target(REST_APP.getHttpListener())
-        .path("ksql")
-        .request(MediaType.APPLICATION_JSON_TYPE)
-        .post(ksqlRequest(
-            "CREATE STREAM " + streamName + " "
-                + "(viewtime bigint, pageid varchar, userid varchar) "
-                + "WITH (kafka_topic='" + PAGE_VIEW_TOPIC + "', value_format='json');"))) {
-
-      assertEquals(Response.Status.OK.getStatusCode(), response.getStatus());
-    } finally {
-      client.close();
-    }
-  }
-
   private static void terminateCluster(final List<String> deleteTopicList) {
     final Client client = TestKsqlRestApp.buildClient();
 
@@ -196,10 +119,6 @@ public class ClusterTerminationTest {
     } finally {
       client.close();
     }
-  }
-
-  private static Entity<?> ksqlRequest(final String sql) {
-    return Entity.json(new KsqlRequest(sql, Collections.emptyMap(), null));
   }
 
   private static Entity<?> terminateClusterRequest(final List<String> deleteTopicList) {
